requirementes.txt
asgiref==3.3.0
Django==3.1.2
django-crispy-forms==1.9.2
djangorestframework==3.12.1
Pillow==8.0.1
pytz==2020.1
sqlparse==0.4.1
django-filter==2.4.0
django-phone-field==1.8.1